stress-cpu-online: remove redundate cast to double

+++ stress-cpu-online.c
@@ -416,10 +416,10 @@ static int stress_cpu_online(stress_args_t *args)
 	}
 	free(cpu_online);
 
-	rate = (offline_count > 0.0) ? (double)offline_duration / offline_count : 0.0;
+	rate = (offline_count > 0.0) ? offline_duration / offline_count : 0.0;
 	stress_metrics_set(args, "millisecs per offline action",
 		rate * STRESS_DBL_MILLISECOND, STRESS_METRIC_HARMONIC_MEAN);
-	rate = (online_count > 0.0) ? (double)online_duration / online_count : 0.0;
+	rate = (online_count > 0.0) ? online_duration / online_count : 0.0;
 	stress_metrics_set(args, "millisecs per online action",
 		rate * STRESS_DBL_MILLISECOND, STRESS_METRIC_HARMONIC_MEAN);
